==> Gandal/AllRegions/RMarkdowns/16_DisGeNET_vs_SFARI.Rmd <==
---
title: 'DisGeNET'
output:
  html_document:
    code_folding: 'hide'
---

```{r load_packages, echo=TRUE, warning=FALSE, message=FALSE}
library(tidyverse) ; library(reshape2) ; library(glue) ; library(plotly) ; library(plotlyutils)
library(RColorBrewer) ; library(viridis) ; require(gridExtra) ; library(GGally) ; library(ggpubr) ; library(rstatix)
library(VennDiagram)
library(expss)
library(biomaRt)
library(disgenet2r)
```

```{r load_data, warning=FALSE, message=FALSE}

# Gandal dataset
load('./../Data/preprocessed_data.RData')
datExpr = datExpr %>% data.frame


# Update gene scores to new SFARI Genes
SFARI_genes = read_csv('./../../../SFARI/Data/SFARI_genes_01-03-2020_w_ensembl_IDs.csv')
SFARI_genes = SFARI_genes[!duplicated(SFARI_genes$ID) & !is.na(SFARI_genes$ID),]


# Add gene symbol
getinfo = c('ensembl_gene_id','external_gene_id')
mart = useMart(biomart='ENSEMBL_MART_ENSEMBL', dataset='hsapiens_gene_ensembl',
               host='feb2014.archive.ensembl.org') ## Gencode v19
gene_names = getBM(attributes=getinfo, filters=c('ensembl_gene_id'), values=rownames(datExpr), mart=mart)

# DisGeNET
DisGeNET = list()
DisGeNET[['asd']] = disease2gene(disease = 'C0004352')@qresult   # Autism Spectrum Disorder
DisGeNET[['scz']] = disease2gene(disease = 'C0036341')@qresult   # Schizophrenia
DisGeNET[['bd']]  = disease2gene(disease = 'C0005586')@qresult   # Bipolar Disorder
DisGeNET[['id']]  = disease2gene(disease = 'C3714756')@qresult   # Intellectual Disability
DisGeNET[['dd']]  = disease2gene(disease = 'C0011581')@qresult   # Depressive Disorder
DisGeNET[['ai']]  = disease2gene(disease = 'C0001973')@qresult   # Alcoholic Intoxication, Chronic
DisGeNET[['bc']]  = disease2gene(disease = 'C0006142')@qresult   # Malignant Carcinoma of Breast
DisGeNET[['cc']]  = disease2gene(disease = 'C0009402')@qresult   # Colorectal Carcinoma
#lc  = disease2gene(disease = 'C0023893')@qresult   # Liver Cirrhosis NO: Only 2 scores
#ht  = disease2gene(disease = 'C0019193')@qresult   # Hepatitis, Toxic No: Only 2 scores
DisGeNET[['dm']]  = disease2gene(disease = 'C0011860')@qresult   # Diabetes Melitus
DisGeNET[['ob']]  = disease2gene(disease = 'C0028754')@qresult   # Obesity
DisGeNET[['ra']]  = disease2gene(disease = 'C0003873')@qresult   # Rheumatoid Arthiritis

SFARI_colour_hue = function(r) {
  pal = c('#FF7631','#FFB100','#E8E328','#b3b3b3','#808080','gray','#d9d9d9')[r]
}


rm(GO_annotations, mart, getinfo, dds, DE_info)
```

## ASD
<br>

There is an ASD dataset in DisGeNET, so I'll check if there is some relation between this scores and SFARI Genes:

SFARI has many more genes than DisGeNET, but they share a big proportion of DisGeNET's genes
```{r fig.width=5}

asd_info = data.frame('ID' = gene_names$ensembl_gene_id, 'gene_symbol' = gene_names$external_gene_id, 
                       'meanExpr' = rowMeans(datExpr)) %>% 
           left_join(SFARI_genes %>% dplyr::select(ID, `gene-score`), by = 'ID') %>%
           left_join(DisGeNET[['asd']] %>% dplyr::select(gene_symbol, score), by = 'gene_symbol') %>%
           mutate('SFARI' = !is.na(`gene-score`), 'DisGeNET' = !is.na(score)) %>% 
           apply_labels('SFARI'= 'SFARI', 'DisGeNET' = 'DisGeNET')

cro_cpct(asd_info$SFARI, list(asd_info$DisGeNET,total()))
  
draw.pairwise.venn(area1 = sum(asd_info$SFARI), area2 =  sum(asd_info$DisGeNET),
                   cross.area = sum(asd_info$SFARI*asd_info$DisGeNET),
                   col = c('#0099cc','#d147a3'), fill = c('#0099cc','#d147a3'), alpha = c(0.3,0.3), 
                   category = c('SFARI','DisGeNET')) %>% grid.draw
```

And both scoring systems agree with each other quite well

```{r, fig.width=5, fig.height=6, fig.align='center'}

wt = asd_info %>% filter(SFARI*DisGeNET == TRUE) %>% mutate(gene_score = `gene-score` %>% as.factor) %>%
     wilcox_test(score~gene_score, p.adjust.method='BH') %>% add_x_position(x = 'group')

increase = 0.03
base = 0.72
pos_y_comparisons = c(base, base + increase, base)

asd_info %>% filter(SFARI*DisGeNET == TRUE) %>% mutate(gene_score = `gene-score` %>% as.factor) %>%
             ggplot(aes(gene_score, score)) + 
             geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = gene_score)) + 
             stat_pvalue_manual(wt, y.position = pos_y_comparisons, tip.length = .02) +
             xlab('SFARI Scores') + ylab('DisGeNET Scores') + 
             scale_fill_manual(values=SFARI_colour_hue(1:3)) + theme_minimal() + theme(legend.position = 'none')

rm(increase, base, pos_y_comparisons, wt)
```
<br><br>

## Mental Disorders
<br>

- Autism Spectrum Disorders (`r nrow(DisGeNET[['asd']])` genes)

- Schizophrenia (`r nrow(DisGeNET[['scz']])`)

- Bipolar Disorder (`r nrow(DisGeNET[['bd']])`)

- Intellectual Disability (`r nrow(DisGeNET[['id']])`)
  
- Depressive Disorder (`r nrow(DisGeNET[['dd']])`)

- Chronic Alcoholic Intoxication (`r nrow(DisGeNET[['ai']])`)

<br>

### DisGeNET scores vs Mean Expression
<br>

There seems to be a weak positive relation between the two variables except for Depressive Disorder and Alcoholism
```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=4}

disgenet_info = data.frame('ID' = gene_names$ensembl_gene_id, 'gene_symbol' = gene_names$external_gene_id, 
                           'meanExpr' = rowMeans(datExpr)) %>% 
                left_join(DisGeNET[['asd']] %>% dplyr::rename('ASD'=score), by = 'gene_symbol') %>%
                left_join(DisGeNET[['scz']] %>% dplyr::rename('Schizophrenia'=score), by = 'gene_symbol') %>%
                left_join(DisGeNET[['bd']] %>% dplyr::rename('BipolarDisorder'=score), by = 'gene_symbol') %>%
                left_join(DisGeNET[['id']] %>% dplyr::rename('IntellectualDisability'=score),by='gene_symbol')%>%
                left_join(DisGeNET[['dd']] %>% dplyr::rename('DepressiveDisorder'=score), by='gene_symbol') %>%
                left_join(DisGeNET[['ai']] %>% dplyr::rename('Alcoholism'=score), by = 'gene_symbol') %>%
                dplyr::select(gene_symbol, meanExpr, ASD, Schizophrenia, BipolarDisorder, 
                              IntellectualDisability, DepressiveDisorder, Alcoholism)


disgenet_info %>% dplyr::select(-gene_symbol) %>% melt(id.vars = 'meanExpr') %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  ggplot(aes(meanExpr, value, color = Disorder)) + geom_point(alpha=0.3) + 
                  facet_grid(~Disorder) + geom_smooth(alpha = 0.2, method = 'loess') + 
                  xlab('Mean Expression') + ylab('Score') + scale_y_log10() + theme_minimal() + 
                  theme(legend.position = 'none')


rm(asd_info)
```

There is a big overlap between SFARI Genes and the genes related to each of the disorders (luckily the overlap is largest with ASD)

```{r, warning=FALSE, message=FALSE}

table_info = disgenet_info %>% 
             mutate('SFARI' = gene_symbol %in% SFARI_genes$`gene-symbol`[!is.na(SFARI_genes$`gene-score`)]) %>%
             dplyr::select(-c(gene_symbol, meanExpr)) %>% melt(id.vars = 'SFARI') %>%
             filter(!is.na(value)) %>% apply_labels(SFARI = 'SFARI', variable = 'DisGeNET')

cro_cpct(table_info$SFARI, list(table_info$variable, total()))

rm(table_info)
```

This overlap could be causing the positive relation between scores and mean expression, to check if this is the case, I'm going to analyse separetly the genes that are in the SFARI list from the rest.

The pattern is quite noisy, but the positive relation seems to come from both SFARI and non-SFARI genes

```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=6}

disgenet_info %>% mutate('SFARI' = gene_symbol %in% SFARI_genes$`gene-symbol`) %>% 
                  dplyr::select(-gene_symbol) %>% melt(id.vars = c('meanExpr', 'SFARI')) %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  ggplot(aes(meanExpr, value, color = SFARI)) + geom_point(alpha=0.3) + 
                  facet_grid(~Disorder) + geom_smooth(alpha = 0.2, method = 'loess') + 
                  xlab('Mean Expression') + ylab('Score') + scale_y_log10() + theme_minimal() + 
                  theme(legend.position = 'bottom')

```

For all disorders except Depressive Disorder, SFARI Genes have significantly higher scores than the rest of the genes

```{r, warning=FALSE, message=FALSE, fig.width=12, fig.height=4}

disgenet_info %>% mutate('SFARI' = gene_symbol %in% SFARI_genes$`gene-symbol`) %>% 
                  dplyr::select(-gene_symbol) %>% melt(id.vars = c('meanExpr', 'SFARI')) %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  ggplot(aes(SFARI, value, fill = SFARI)) + 
                  geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3) + 
                  facet_grid(~Disorder) + scale_y_log10() +
                  stat_compare_means(label = 'p.signif', method = 't.test', 
                                     method.args = list(var.equal = FALSE)) +
                  xlab('Belong to SFARI Genes') + ylab('DisGeNET Score') +
                  theme_minimal() + theme(legend.position = 'bottom')


plot_data = disgenet_info %>% left_join(SFARI_genes %>% dplyr::select(`gene-symbol`, `gene-score`), 
                            by = c('gene_symbol'='gene-symbol')) %>% 
                  mutate(SFARI = ifelse(is.na(`gene-score`), 'non-SFARI', `gene-score` %>% as.character)) %>%
                  mutate(SFARI = factor(SFARI, levels = c('1','2','3','non-SFARI'))) %>%
                  dplyr::select(-c(gene_symbol, `gene-score`)) %>%
                  melt(id.vars = c('meanExpr', 'SFARI')) %>% dplyr::rename('Disorder' = variable) %>%
                  filter(!is.na(value))

wt = plot_data %>% group_by(Disorder) %>% wilcox_test(value~SFARI, p.adjust.method='BH') %>% 
     add_xy_position(x = 'SFARI')

increase = 0.05
base = 0.72
pos_y_comparisons = c(base + c(0,1,3)*increase, base+c(0,2)*increase, base)

plot_data %>% ggplot(aes(SFARI, value)) + 
                  geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = SFARI)) + 
                  facet_grid(~Disorder) + 
                  stat_pvalue_manual(wt, y.position = pos_y_comparisons, tip.length = .01) +
                  xlab('SFARI Gene Score') + ylab('DisGeNET Score') + #scale_y_log10() +
                  scale_fill_manual(values = SFARI_colour_hue(c(1:3,6))) +
                  theme_minimal() + theme(legend.position = 'bottom')


rm(asd_info, comparisons, increase, base, pos_y_comparisons, wt)
```
<br>

### DisGeNET binned scores vs Mean Expression
<br>

The relation is sometimes not that clear in the continuous scale (maybe because there aren't that many points and the relation is not that strong?) I'm going to separate the scores into three groups to see if a pattern is more visible this way

Binned scores (selected to balance the number of genes in each group): 

- Low: (0,0.31]

- Medium: (0.31,0.45]

- High: (0.45,1]

<br>

Same conclusions as before, just different visualisations
<br>

```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=4}

plot_data = disgenet_info %>% dplyr::select(-gene_symbol) %>% melt(id.vars = 'meanExpr') %>% 
            dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
            mutate(binned_scores = cut(value, breaks = c(0, 0.31, 0.45, 1), 
                                       labels = c('Low','Medium','High'))) %>%
            mutate(binned_scores = factor(binned_scores, levels = c('High','Medium','Low')))

wt = plot_data %>% group_by(Disorder) %>% wilcox_test(meanExpr~binned_scores, p.adjust.method='BH') %>% 
     add_xy_position(x = 'binned_scores')

increase = 0.9
base = 16
pos_y_comparisons =c(base, base+increase, base)

plot_data %>% ggplot(aes(binned_scores, meanExpr)) + facet_grid(~Disorder) +
                  geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = Disorder)) + 
                  stat_pvalue_manual(wt, y.position = pos_y_comparisons, tip.length = .01) +
                  xlab('Binned Scores') + ylab('Mean Expression') + 
                  theme_minimal() + theme(legend.position = 'none')


rm(asd_info, increase, base, pos_y_comparisons)
```
<br>

Separating SFARI Genes
```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=6}

plot_data = disgenet_info %>% mutate('SFARI' = gene_symbol %in% SFARI_genes$`gene-symbol`) %>% 
            dplyr::select(-gene_symbol) %>% melt(id.vars = c('meanExpr', 'SFARI')) %>% 
            dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
            mutate(binned_scores = cut(value, breaks = c(0, 0.31, 0.45, 1), 
                                       labels = c('Low','Medium','High'))) %>%
            mutate(binned_scores = factor(binned_scores, levels = c('High','Medium','Low')))

wt = plot_data %>% group_by(Disorder, SFARI) %>% wilcox_test(meanExpr~binned_scores, p.adjust.method='BH') %>% 
     add_xy_position(x = 'binned_scores')

increase = 1
base = 16
pos_y_comparisons = c( c(base, base+increase) ,rep(c(base, base+increase, base),10), c(base, base+increase) )

plot_data %>% ggplot(aes(binned_scores, meanExpr)) + facet_grid(SFARI ~ Disorder) +
              geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = Disorder)) + 
              stat_pvalue_manual(wt, y.position = pos_y_comparisons, tip.length = .01) +
              xlab('Binned Scores') + ylab('Mean Expression') +
              theme_minimal() + theme(legend.position = 'none')


rm(comparisons, increase, base, pos_y_comparisons)
```
<br><br>

---

<br><br>

## Other Diseases
<br>

- Malignant neoplasm of breast (1074)

- Experimental Liver Cirrhosis (774)

- Colorectal carcinoma (702)

- Hepatitis, Toxic (404)

- Diabetes Melitus (221)

<br>

### DisGeNET scores vs Mean Expression
<br>

There doesn't seem to be a recognisable relation between mean expression and score for any of these diseases

```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=4}

disgenet_info = data.frame('ID' = gene_names$ensembl_gene_id, 'gene_symbol' = gene_names$external_gene_id, 
                       'meanExpr' = rowMeans(datExpr)) %>% 
                left_join(DisGeNET[['bc']] %>% dplyr::rename('BreastCancer'=score), by = 'gene_symbol') %>%
                left_join(DisGeNET[['cc']] %>% dplyr::rename('ColorectalCarcinoma'=score), by='gene_symbol')%>%
                left_join(DisGeNET[['dm']] %>% dplyr::rename('Diabetes'=score), by = 'gene_symbol') %>%
                left_join(DisGeNET[['ob']] %>% dplyr::rename('Obesity'=score), by = 'gene_symbol') %>%
                left_join(DisGeNET[['ra']] %>% dplyr::rename('Arthritis'=score), by = 'gene_symbol') %>%
                dplyr::select(gene_symbol, meanExpr, BreastCancer, ColorectalCarcinoma, Diabetes, Obesity,
                              Arthritis)


disgenet_info %>% dplyr::select(-gene_symbol) %>% melt(id.vars = 'meanExpr') %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  ggplot(aes(meanExpr, value, color = Disorder)) + geom_point(alpha=0.3) + 
                  facet_grid(~Disorder) + geom_smooth(alpha = 0.2, method = 'loess') + 
                  xlab('Mean Expression') + ylab('Score') + scale_y_log10() + theme_minimal() + 
                  theme(legend.position = 'none')

```

There is still an overlap with SFARI Genes, not as big as with neurological disorders, but still weird, since these diseases aren't related to ASD

```{r, warning=FALSE, message=FALSE}

table_info = disgenet_info %>% 
             mutate('SFARI' = gene_symbol %in% SFARI_genes$`gene-symbol`[!is.na(SFARI_genes$`gene-score`)]) %>%
             dplyr::select(-c(gene_symbol, meanExpr)) %>% melt(id.vars = 'SFARI') %>%
             filter(!is.na(value)) %>% apply_labels(SFARI = 'SFARI', variable = 'DisGeNET')

cro_cpct(table_info$SFARI, list(table_info$variable, total()))

rm(table_info)
```

This seems to be just noise, perhaps there aren't enough points for a reliable fit

```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=6}

disgenet_info %>% mutate('SFARI' = gene_symbol %in% SFARI_genes$`gene-symbol`) %>% 
                  dplyr::select(-gene_symbol) %>% melt(id.vars = c('meanExpr', 'SFARI')) %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  ggplot(aes(meanExpr, value, color = SFARI)) + geom_point(alpha=0.3) + 
                  facet_grid(~Disorder) + geom_smooth(alpha = 0.2, method = 'loess') + 
                  xlab('Mean Expression') + ylab('Score') + scale_y_log10() + theme_minimal() + 
                  theme(legend.position = 'bottom')

```

There is no longer a relation between SFARI Genes and the scores of this diseases, either all together or separating them by scores

```{r, warning=FALSE, message=FALSE, fig.width=12, fig.height=4}

disgenet_info %>% mutate('SFARI' = gene_symbol %in% SFARI_genes$`gene-symbol`) %>% 
                  dplyr::select(-gene_symbol) %>% melt(id.vars = c('meanExpr', 'SFARI')) %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  ggplot(aes(SFARI, value, fill = SFARI)) + 
                  geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3) + 
                  facet_grid(~Disorder) + scale_y_log10() + ggtitle('DisGeNET Scores vs SFARI Genes') +
                  stat_compare_means(label = 'p.signif', method = 't.test', 
                                     method.args = list(var.equal = FALSE)) +
                  xlab('Belong to SFARI Genes') + ylab('DisGeNET Score') +
                  theme_minimal() + theme(legend.position = 'bottom')

plot_data = disgenet_info %>% left_join(SFARI_genes %>% dplyr::select(`gene-symbol`, `gene-score`), 
                            by = c('gene_symbol'='gene-symbol')) %>% 
                  mutate(SFARI = ifelse(is.na(`gene-score`), 'non-SFARI', `gene-score` %>% as.character)) %>%
                  mutate(SFARI = factor(SFARI, levels = c('1','2','3','non-SFARI'))) %>%
                  dplyr::select(-c(gene_symbol, `gene-score`)) %>%
                  melt(id.vars = c('meanExpr', 'SFARI')) %>% dplyr::rename('Disorder' = variable) %>%
                  filter(!is.na(value))

wt = plot_data %>% group_by(Disorder) %>% wilcox_test(value~SFARI, p.adjust.method='BH') %>% 
     add_xy_position(x = 'SFARI')

increase = 0.05
base = 0.8
pos_y_comparisons = c( rep(c(base + c(0,1,3)*increase, base+c(0,2)*increase, base),4), c(base, base+increase, base))

plot_data %>% ggplot(aes(SFARI, value)) + 
                  geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = SFARI)) + 
                  facet_grid(~Disorder) + 
                  stat_pvalue_manual(wt, y.position = pos_y_comparisons, tip.length = .01) +
                  xlab('SFARI Gene Score') + ylab('DisGeNET Score') + #scale_y_log10() +
                  scale_fill_manual(values = SFARI_colour_hue(c(1:3,6))) +
                  theme_minimal() + theme(legend.position = 'bottom')


rm(asd_info, comparisons, increase, base, pos_y_comparisons)
```
<br>

### DisGeNET binned scores vs Mean Expression
<br>

Usin the same binned scores as before: 

- Low: (0,0.31]

- Medium: (0.31,0.45]

- High: (0.45,1]

<br>

Same conclusions as before, just different visualisations
<br>

```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=4}

plot_data = disgenet_info %>% dplyr::select(-gene_symbol) %>% melt(id.vars = 'meanExpr') %>% 
            dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
            mutate(binned_scores = cut(value, breaks = c(0, 0.31, 0.45, 1), 
                                       labels = c('Low','Medium','High'))) %>%
            mutate(binned_scores = factor(binned_scores, levels = c('High','Medium','Low')))

wt = plot_data %>% group_by(Disorder) %>% wilcox_test(meanExpr~binned_scores, p.adjust.method='BH') %>% 
     add_xy_position(x = 'binned_scores')

increase = 0.9
base = 15
pos_y_comparisons =c(base, base+increase, base)

plot_data %>% ggplot(aes(binned_scores, meanExpr)) + facet_grid(~Disorder) +
                  geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = Disorder)) + 
                  stat_pvalue_manual(wt, y.position = pos_y_comparisons, tip.length = .01) +
                  xlab('Binned Scores') + ylab('Mean Expression') + 
                  theme_minimal() + theme(legend.position = 'none')


rm(asd_info, increase, base, pos_y_comparisons)
```
<br>

Separating SFARI Genes
```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=6}

plot_data = disgenet_info %>% mutate('SFARI' = gene_symbol %in% SFARI_genes$`gene-symbol`) %>% 
            dplyr::select(-gene_symbol) %>% melt(id.vars = c('meanExpr', 'SFARI')) %>% 
            dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
            mutate(binned_scores = cut(value, breaks = c(0, 0.31, 0.45, 1), 
                                       labels = c('Low','Medium','High'))) %>%
            mutate(binned_scores = factor(binned_scores, levels = c('High','Medium','Low')))

wt = plot_data %>% group_by(Disorder, SFARI) %>% wilcox_test(meanExpr~binned_scores, p.adjust.method='BH') %>% 
     add_xy_position(x = 'binned_scores')

increase = 1
base = 15
pos_y_comparisons = c(base, base+increase, base)

plot_data %>% ggplot(aes(binned_scores, meanExpr)) + facet_grid(SFARI ~ Disorder) +
              geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = Disorder)) + 
              stat_pvalue_manual(wt, y.position = pos_y_comparisons, tip.length = .01) +
              xlab('Binned Scores') + ylab('Mean Expression') +
              theme_minimal() + theme(legend.position = 'none')


rm(comparisons, increase, base, pos_y_comparisons)

```
<br><br>

---

<br><br>

## Schizophrenia
<br>

It seems that this is the only disorder with a significant relation between its scores and mean expression, so I'll try to see if this relation comes from the SFARI Genes or ASD genes
<br>

### Schizophrenia vs SFARI Genes
<br>

The relation of both scoring systems seems to be stronger when they don't share genes with each other, but it probably is just because there are less observations in the intersection than in the other two regions of the Venn Diagram, which makes the results less robust
```{r, fig.width=5, fig.align = 'center'}

scz_SFARI = data.frame('ID' = gene_names$ensembl_gene_id, 'gene_symbol' = gene_names$external_gene_id, 
                       'meanExpr' = rowMeans(datExpr)) %>% 
            left_join(DisGeNET[['scz']] %>% dplyr::rename('scz'=score), by = 'gene_symbol') %>%
            dplyr::select(gene_symbol, meanExpr, scz) %>%
            full_join(SFARI_genes %>% filter(!is.na(`gene-score`)) %>% dplyr::rename('SFARI' = `gene-score`) %>%
                        dplyr::select(`gene-symbol`, SFARI), by = c('gene_symbol'='gene-symbol')) %>%
            mutate('in_scz' = !is.na(scz), 'in_SFARI' = !is.na(SFARI))


draw.pairwise.venn(area1 = sum(scz_SFARI$in_scz), area2 = sum(scz_SFARI$in_SFARI),
                   cross.area = sum(scz_SFARI$in_scz*scz_SFARI$in_SFARI),
                   col = c('#ffcc00','#0099cc'), fill = c('#ffcc00','#0099cc'), alpha = c(0.3,0.3), 
                   category = c('Scz DisGeNET','SFARI')) %>% grid.draw

```

```{r warning=FALSE, message=FALSE, fig.width=10, fig.height = 4}
  
# SCZ with and without SFARI
plot_data = scz_SFARI %>% filter(!is.na(scz)) %>% dplyr::select(meanExpr, in_SFARI, scz) %>% 
            melt(id.vars = c('meanExpr','in_SFARI')) %>% 
            mutate(in_SFARI = factor(in_SFARI, levels = c(TRUE, FALSE))) %>%
            mutate(binned_scores = cut(value, breaks = c(0, 0.31, 0.45, 1), 
                                       labels = c('Low','Medium','High'))) %>%
            mutate(binned_scores = factor(binned_scores, levels = c('High','Medium','Low')))

wt = plot_data %>% group_by(in_SFARI) %>% wilcox_test(meanExpr~binned_scores, p.adjust.method='BH') %>% 
     add_xy_position(x = 'binned_scores')

increase = 1
base = 16
pos_y_comparisons = c(base, base+increase, base)

plot_1 = plot_data %>%
         ggplot(aes(binned_scores, meanExpr)) + facet_wrap(~in_SFARI) +
         geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = in_SFARI)) + 
         stat_pvalue_manual(wt, y.position = pos_y_comparisons, tip.length = .01) +
         xlab('Binned Scores') + ylab('Mean Expression') + 
         theme_minimal() + theme(legend.position = 'none') + ggtitle('Schizophrenia divided by SFARI')


# SFARI with and without SCZ
plot_data = scz_SFARI %>% filter(!is.na(SFARI)) %>% dplyr::select(meanExpr, in_scz, SFARI) %>% 
            melt(id.vars = c('meanExpr','in_scz'))
  
wt = plot_data %>% group_by(in_scz) %>% mutate(value = value %>% as.factor) %>% 
     wilcox_test(meanExpr~value, p.adjust.method='BH') %>% add_xy_position(x = 'value')

increase = 0.9
base = 15
pos_y_comparisons = c(base, base+increase, base)

plot_2 = plot_data %>%
         ggplot(aes(value %>% as.factor, meanExpr)) + facet_wrap(~in_scz) +
         geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = as.factor(value))) + 
         stat_pvalue_manual(wt, y.position = pos_y_comparisons, tip.length = .01) +
         scale_fill_manual(values = SFARI_colour_hue(1:3)) +
         xlab('SFARI Scores') + ylab('Mean Expression') + 
         theme_minimal() + theme(legend.position = 'none') + ggtitle('SFARI divided by Schizophrenia')


grid.arrange(plot_2, plot_1, nrow=1)

rm(plot_1, plot_2, scz_SFARI, wt)
```

### Schizophrenia vs ASD DisGeNET Genes
<br>

```{r, fig.width=5, fig.align = 'center'}

scz_asd = data.frame('ID' = gene_names$ensembl_gene_id, 'gene_symbol' = gene_names$external_gene_id, 
                       'meanExpr' = rowMeans(datExpr)) %>% 
            left_join(DisGeNET[['scz']] %>% dplyr::rename('scz'=score), by = 'gene_symbol') %>%
            left_join(DisGeNET[['asd']] %>% dplyr::rename('asd'=score), by = 'gene_symbol') %>%  
            dplyr::select(gene_symbol, meanExpr, scz, asd) %>%
            mutate('in_scz' = !is.na(scz), 'in_asd' = !is.na(asd))


draw.pairwise.venn(area1 = sum(scz_asd$in_scz), area2 = sum(scz_asd$in_asd),
                   cross.area = sum(scz_asd$in_scz*scz_asd$in_asd),
                   col = c('#ffcc00','#d147a3'), fill = c('#ffcc00','#d147a3'), alpha = c(0.3,0.3), 
                   category = c('Scz DisGeNET','ASD DisGeNET')) %>% grid.draw

```

```{r warning=FALSE, message=FALSE, fig.width=10, fig.height = 4}

# SCZ with and without ASD
plot_data = scz_asd %>% filter(!is.na(scz)) %>% dplyr::select(meanExpr, in_asd, scz) %>% 
            melt(id.vars = c('meanExpr','in_asd')) %>% 
            mutate(binned_scores = cut(value, breaks = c(0, 0.31, 0.45, 1), 
                                    labels = c('Low','Medium','High'))) %>%
            mutate(binned_scores = factor(binned_scores, levels = c('High','Medium','Low')))

wt = plot_data %>% group_by(in_asd) %>% wilcox_test(meanExpr~binned_scores, p.adjust.method='BH') %>% 
     add_xy_position(x = 'binned_scores')

increase = 1
base = 16
pos_y_comparisons = c(base, base+increase, base)

plot_1 = plot_data %>%
         ggplot(aes(binned_scores, meanExpr)) + facet_wrap(~in_asd) +
         geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = in_asd)) + 
         stat_pvalue_manual(wt, y.position = pos_y_comparisons, tip.length = .01) +
         xlab('Binned Scores') + ylab('Mean Expression') + 
         theme_minimal() + theme(legend.position = 'none') + ggtitle('Schizophrenia divided by ASD')


# ASD with and without SCZ
plot_data = scz_asd %>% filter(!is.na(asd)) %>% dplyr::select(meanExpr, in_scz, asd) %>% 
            melt(id.vars = c('meanExpr','in_scz')) %>% 
            mutate(in_scz = factor(in_scz, levels = c(TRUE, FALSE))) %>%
            mutate(binned_scores = cut(value, breaks = c(0, 0.31, 0.45, 1), 
                                       labels = c('Low','Medium','High'))) %>%
            mutate(binned_scores = factor(binned_scores, levels = c('High','Medium','Low')))

wt = plot_data %>% group_by(in_scz) %>% wilcox_test(meanExpr~binned_scores, p.adjust.method='BH') %>% 
     add_xy_position(x = 'binned_scores')

plot_2 = plot_data %>%
         ggplot(aes(binned_scores, meanExpr)) + facet_wrap(~in_scz) +
         geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = in_scz)) + 
         stat_pvalue_manual(wt, y.position = pos_y_comparisons, tip.length = .01) +
         xlab('SFARI Scores') + ylab('Mean Expression') + 
         theme_minimal() + theme(legend.position = 'none') + ggtitle('ASD divided by Schizophrenia')


grid.arrange(plot_1, plot_2, nrow=1)

rm(plot_1, plot_2, scz_asd, wt)
```
<br><br>

### Schizophrenia vs ASD (DisGeNET + SFARI)
<br>

Just to be sure to cover all the ASD-related genes I'm going to combine the DisGeNET ASD lit with SFARI.

This time we cannot do the boxplots corresponding to ASD because we have two different scoring systems combined

```{r, fig.width=5, fig.align = 'center'}

scz_asd = data.frame('ID' = gene_names$ensembl_gene_id, 'gene_symbol' = gene_names$external_gene_id, 
                       'meanExpr' = rowMeans(datExpr)) %>% 
          left_join(DisGeNET[['scz']] %>% dplyr::rename('scz'=score), by = 'gene_symbol') %>%
          left_join(DisGeNET[['asd']] %>% dplyr::rename('asd'=score), by = 'gene_symbol') %>% 
          left_join(SFARI_genes %>% filter(!is.na(`gene-score`)) %>% dplyr::rename('SFARI' = `gene-score`) %>%
                    dplyr::select(`gene-symbol`, SFARI), by = c('gene_symbol'='gene-symbol')) %>%
          dplyr::select(gene_symbol, meanExpr, scz, asd, SFARI) %>%
          mutate('in_scz' = !is.na(scz), 'in_asd' = !is.na(asd) | !is.na(SFARI))


draw.pairwise.venn(area1 = sum(scz_asd$in_scz), area2 = sum(scz_asd$in_asd),
                   cross.area = sum(scz_asd$in_scz*scz_asd$in_asd),
                   col = c('#ffcc00','#9900cc'), fill = c('#ffcc00','#9900cc'), alpha = c(0.3,0.3), 
                   category = c('Scz DisGeNET','ASD')) %>% grid.draw

```

```{r warning=FALSE, message=FALSE, fig.width=5, fig.height = 4, fig.align='right'}

plot_data = scz_asd %>% filter(!is.na(scz)) %>% dplyr::select(meanExpr, in_asd, scz) %>% 
            melt(id.vars = c('meanExpr','in_asd')) %>% 
            mutate(binned_scores = cut(value, breaks = c(0, 0.31, 0.45, 1), 
                                       labels = c('Low','Medium','High'))) %>%
            mutate(binned_scores = factor(binned_scores, levels = c('High','Medium','Low'))) %>%
            mutate(in_asd = factor(in_asd, levels = c(TRUE, FALSE)))

wt = plot_data %>% group_by(in_asd) %>% wilcox_test(meanExpr~binned_scores, p.adjust.method='BH') %>% 
     add_xy_position(x = 'binned_scores')
  
increase = 1
base = 15
pos_y_comparisons = c(base, base+increase, base)

# SCZ with and without ASD
plot_data %>% ggplot(aes(binned_scores, meanExpr)) + facet_wrap(~in_asd) +
            geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = in_asd)) + 
            stat_pvalue_manual(wt, y.position = pos_y_comparisons, tip.length = .01) +
            xlab('Binned Scores') + ylab('Mean Expression') + 
            theme_minimal() + theme(legend.position = 'none') + ggtitle('Schizophrenia divided by ASD')


rm(comparisons, increase, base, pos_y_comparisons, scz_asd)
```
<br><br>

---

<br><br>

## Conclusions
<br>

- SFARI and DisGeNET ASD scores agree between them, both in membership as in scores

- There is a big overlap between SFARI Genes and the genes in all of the diseases studied, considering that there are only 789 SFARI Genes

- The closer the disorder is to ASD:

    - The stronger the relation between SFARI Genes and the DisGeNET scores 
    
    - The stronger the relation between score and mean expression
    
    - Schizophrenia having statistical significant differences, Bipolar Disorder and Intellectual Disability visible diferences but not strong enough to be significant, and the rest of the diseases not having any recognisable pattern
    
- The Bias related to mean level of expression found in Schizophrenia-related genes doesn't seem to be related to either genes from the DisGeNET ASD list nor to the SFARI genes, but intrinsic to Schizophrenia (and the reverse is also true, the bias found in the SFARI Genes doesn't come from schizophrenia-related genes)

<br><br>
